Revert dependabot updates and re-enable package-lock
I am not sure why the previous team(s) decided to use manual updates and disabled package locks, but this will definitely be changed. I reverted the dependabot merges that the professor did to bring the application back to the original state the previous teams left it in for testing.