Install electron.js
Browser inputs are not allowed to see the full path of the file being selected. Luckily, electron,js is, so we have to switch to that to be able to pass file paths to the docker without uploading the actual file. Uploading the file would duplicate the data on disk which is not an acceptable solution.